working on articles...

+++ Linx_node/src/routes/routingAccount.js
@@ -29,7 +29,6 @@ const storage = multer.diskStorage({
     filename: function (req, file, cb) {
         const artid = req.body.articleid;
         const userDirectory = path.join('C:/Users/cunns/Documents/TFGLinx/articles', req.params.userid);
-        if(file){
             fs.readdir(userDirectory, (err, files) => {
                 if (err) {
                     console.error('Error al leer el directorio:', err);
@@ -59,48 +58,53 @@ const storage = multer.diskStorage({
                     cb(null, file.originalname);
                 }
             });
-        }
-
+        
     }
+
 })
 
 const upload = multer({
     storage: storage
 })
 
-const deleteFile = (req, res) => {
-    const { userid, articleid } = req.params;
+const deleteFile = (req, res, next) => {
+    const { userid } = req.params;
+    const filename = req.query.img;
     const userDirectory = path.join('C:/Users/cunns/Documents/TFGLinx/articles', userid);
 
-    fs.readdir(userDirectory, (err, files) => {
-        if (err) {
-            console.error('Error al leer el directorio:', err);
-            return res.status(500).json({ message: 'Error al leer el directorio' });
-        }
-
-        let fileToDelete = null;
-        for (const filename of files) {
-            const [existingArtid] = filename.split('__');
-            if (existingArtid === articleid) {
-                fileToDelete = filename;
-                break;
+    if(filename){
+        fs.readdir(userDirectory, (err, files) => {
+            if (err) {
+                console.error('Error al leer el directorio:', err);
+                return res.status(500).json({ message: 'Error al leer el directorio' });
             }
-        }
-
-        if (fileToDelete) {
-            const filePath = path.join(userDirectory, fileToDelete);
-            fs.unlink(filePath, (err) => {
-                if (err) {
-                    console.error('Error al eliminar el archivo:', err);
-                    return res.status(500).json({ message: 'Error al eliminar el archivo' });
-                } else {
-                    return res.status(200).json({ message: 'Archivo eliminado exitosamente' });
+    
+            let fileToDelete = null;
+            for (const name of files) {
+                if (name === filename) {
+                    fileToDelete = filename;
+                    break;
                 }
-            });
-        } else {
-            return res.status(404).json({ message: 'Archivo no encontrado' });
-        }
-    });
+            }
+    
+            if (fileToDelete) {
+                const filePath = path.join(userDirectory, fileToDelete);
+                fs.unlink(filePath, (err) => {
+                    if (err) {
+                        console.error('Error al eliminar el archivo:', err);
+                        return res.status(500).json({ message: 'Error al eliminar el archivo' });
+                    } else {
+                        return res.status(200).json({ message: 'Archivo eliminado exitosamente' });
+                    }
+                });
+            } else {
+                return res.status(404).json({ message: 'Archivo no encontrado' });
+            }
+        });
+    }else{
+        next();
+    }
+    
 };
 
 const jwt = require('jsonwebtoken');
@@ -127,6 +131,7 @@ async function checkJWT(req, res, next) {
             );
     }
 }
+
 router.use('/uploads', express.static(path.join('C:/Users/cunns/Documents/TFGLinx/articles/')))
 
 router.get('/trackLocationGeocode', AccountController.trackLocationGeocode);
@@ -137,7 +142,7 @@ router.put('/:userid', checkJWT, AccountController.modifyAccountData);
 router.post('/resetPwd', checkJWT, AccountController.resetPassword);
 router.get('/activate_account', AccountController.activateAccount);
 router.put('/chat/:roomkey', AccountController.storeChatMessage);
-router.post('/:userid/article', upload.single('file'), AccountController.newArticle);
+router.post('/:userid/article',upload.single('file'), AccountController.newArticle);
 router.put('/:userid/article/:artid', upload.single('file'), AccountController.editArticle);
 router.delete('/:userid/article/:artid',deleteFile, AccountController.deleteArticle)
 router.get('/:userid/chat/:linxuserid', AccountController.getChats)
